rename prime-nums.hpp header
